\name{import.psmc}
\alias{import.psmc}
\title{Import PSMC Files}
\description{
  Imports file output by the PSMC application.
}
\usage{
import.psmc(file)
}
\arguments{
  \item{file}{a character string giving a file name; can be a connection.}
}
\value{an object of class \code{"psmc"}.}
\author{Emmanuel Paradis}
\seealso{\code{\link{psmc}}, \code{\link{plot.psmc}}}
\keyword{IO}
